refactoring + added logic
