+++ CS02.12.24.sln
@@ -1,13 +1,14 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35521.163 d17.12
+VisualStudioVersion = 17.12.35521.163
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CS02.12.24", "CS02.12.24.csproj", "{4E49FA37-CA14-45B9-88B5-032409385A5A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Models", "Models", "{79B7E0AF-934B-4DCD-BEE0-409D5B81BADC}"
 	ProjectSection(SolutionItems) = preProject
 		FinanceManager.cs = FinanceManager.cs
+		Run.cs = Run.cs
 		Transaction.cs = Transaction.cs
 	EndProjectSection
 EndProject
